mpv: add testing profile for prescalers
Still a WIP, but the shaders should most likely work.

The first two parameters are scale_x, scale_y, the scale factor in both
direction for this pass.

The next two parameters indicate which pixel from the upscaled
[0, scale_x)x[0,scale_y) subregion is adopt directly from the original pixel.
This parameter is used to calculate offsets.

The last parameter is used as "pass" macro.

Changed the interface of superxbr, in order to align with the nnedi3 shader.
These shaders are supposed to be applied on luma plane only for YUV source,
otherwise to all RGBA planes after colorspace conversion.

Note that this interface is just for prototype testing, there is no
chance to get it merged in this way.